Move queries to configuration
