--- config/prometheusqueries.sample.json
@@ -1,20 +1,0 @@
-[
-    {
-        "Description": "Request duration",
-        "Query": "rate(mattermost_http_request_duration_seconds_sum[5m])/rate(mattermost_http_request_duration_seconds_count[5m])",
-        "Threshold": 1.0,
-        "Alert": true
-    },
-    {
-        "Description": "Total amount of websockets",
-        "Query": "mattermost_http_websockets_total",
-        "Threshold": 0.0,
-        "Alert": false
-    },
-    {
-        "Description": "Average database store time",
-        "Query": "avg(rate(mattermost_db_store_time_sum[5m]))",
-        "Threshold": 0.05,
-        "Alert": true
-    }
-]
